feat: enable multi-core scheduler for rp2350
