+++ src/runtime/runtime_rp2350.go
@@ -3,84 +3,14 @@
 package runtime
 
 import (
-	"device/arm"
-	"machine"
-	"machine/usb/cdc"
+	"device/rp"
 )
 
-// machineTicks is provided by package machine.
-func machineTicks() uint64
-
-// machineLightSleep is provided by package machine.
-func machineLightSleep(uint64)
-
-// ticks returns the number of ticks (microseconds) elapsed since power up.
-func ticks() timeUnit {
-	t := machineTicks()
-	return timeUnit(t)
-}
-
-func ticksToNanoseconds(ticks timeUnit) int64 {
-	return int64(ticks) * 1000
-}
-
-func nanosecondsToTicks(ns int64) timeUnit {
-	return timeUnit(ns / 1000)
-}
-
-func sleepTicks(d timeUnit) {
-	if d <= 0 {
-		return
-	}
-
-	if hasScheduler {
-		// With scheduler, sleepTicks may return early if an interrupt or
-		// event fires - so scheduler can schedule any go routines now
-		// eligible to run
-		machineLightSleep(uint64(d))
-		return
-	}
-
-	// Busy loop
-	sleepUntil := ticks() + d
-	for ticks() < sleepUntil {
-	}
-}
-
-func waitForEvents() {
-	arm.Asm("wfe")
-}
-
-func putchar(c byte) {
-	machine.Serial.WriteByte(c)
-}
-
-func getchar() byte {
-	for machine.Serial.Buffered() == 0 {
-		Gosched()
-	}
-	v, _ := machine.Serial.ReadByte()
-	return v
-}
-
-func buffered() int {
-	return machine.Serial.Buffered()
-}
-
-// machineInit is provided by package machine.
-func machineInit()
-
-func init() {
-	machineInit()
-
-	cdc.EnableUSBCDC()
-	machine.USBDev.Configure(machine.UARTConfig{})
-	machine.InitSerial()
-}
-
-//export Reset_Handler
-func main() {
-	preinit()
-	run()
-	exit(0)
-}
+const (
+	// On RP2040 each core has a different IRQ number: SIO_IRQ_PROC0 and SIO_IRQ_PROC1.
+	// On RP2350 both cores share the same irq number (SIO_IRQ_PROC) just with a
+	// different SIO interrupt output routed to that IRQ input on each core.
+	// https://www.raspberrypi.com/documentation/pico-sdk/high_level.html#group_pico_multicore_1ga1413ebfa65114c6f408f4675897ac5ee
+	sioIrqFifoProc0 = rp.IRQ_SIO_IRQ_FIFO
+	sioIrqFifoProc1 = rp.IRQ_SIO_IRQ_FIFO
+)
